Improve config_path resolution
Previously we used passwd to determine the users home directory then
forced then appened /.config/ to that path. This was very unflexable and
didn't allow the user any chance to change it. with these changes we now
try to find a user specified config dir before falling back to the
passwd solution.

As of now we are checking the following:

XDG_CONFIG_HOME
HOME
passwd

Signed-off-by: Chris Frank <chris@cfrank.org>