natwm: Move to use xcb_wait_for_event
Instead of using poll_for_event which would be better used if we were
listening to the socket, we should block the wm_thread until a new event
comes through. This is fine since the thread is only used for handling
events

With this we need to remove RELOAD support for now since our outer loop
is only checking for RUNNING, and in order to listen to RELOAD we would
need to have a second loop. This functionality will be reintroduced when
we implement ipc calls

Also there was some general cleanup