tree: add non_primitive tests for remove
This adds some initial tests related to adding and removing non
primitive type into the tree. We make sure that all free, and compare
functions are called, and that the removal is performed correctly for
the root node. More tests will come in a future commit.

This commit also adds a method for finding a leaf sibling.

The take away from this commit is not to send stack pointers to heap
allocated pointers to tree_insert...