monitor: Rework monitor setup
Before we were planning on just storing monitor information as rects and
not worrying about them. But when thinking about RANDR updates and
monitor additions and removals the only solution to those problems with
the rects approach is a full tear down of the monitor list. With this
new approach some id info is stored with the monitor. That will make
additions and delitions as well as property changes easier, and more
efficient

Signed-off-by: Chris Frank <chris@cfrank.org>