monitor: store extension with monitor list
at the very least we are going to need to keep track of the extension
data for listening to RANDR events, but I anticipate needing it for
other things. For that reason I am now storing the extension data (type
and cache) with the monitor list inside a new struct.

Also there was some general cleanup