core/tile: Initial implementation of basic tiles
Tiles are the initial name of the client containers. These tiles will
handle managing the size of windows in both their tiled and floating
states and will also handle the tiling of windows.

Each workspace will have a tree of tiles, and each tile will have a
group of clients.

This is a rough implementation, but it's looking pretty cool

Signed-off-by: Chris Frank <chris@cfrank.org>