core/theme: Better error handling and more consistant returns
Log / return better warnings when we can't find the config array we are
looking for.

Instead of returning a fully default border_theme when the length is
incorrect we should do something more similar to color_theme and return
an error and leave the passed border theme alone. This makes it possible
to depend on the error return result without having to also check if the
passed theme value is NULL