common/string: Fix strncpy warning
This fixes a warning that was being thrown because gcc throught we were
not adding the null terminating byte to the resulting string. We
actually were but strncpy did not know that the buffer was actually
strlen + 1 since we were only writing strlen bytes and then after the
call appending the \0.

This commit does thing the correct way and gives strncpy enough buffer
space to add the \0 itself