core/client: Initial implementation of client
We are starting to move toward a floating implementation as I feel this
will be better for the project and more interesting for me to develop.

This was brought on by my boredom in writing tree code. I would rather
explore XCB then write a million tree traversal functions - especially
in C

Signed-off-by: Chris Frank <chris@cfrank.org>