core/workspace: Attempt to find correct window to focus after destroying
When destroying a window that we are not managing sometimes it will send
focus back to NONE or ROOT. Usually if there is another window active on
the visible workspace we want to focus back on that. At the very least
we want to make sure that the root window is properly focused.

Signed-off-by: Chris Frank <chris@cfrank.org>