core/client: Queue button index 1 events until after focus
We need to first queue the button index 1 event and wait until we are
finished focusing the workspace (if needed) and the client. Once we have
done this we can release the event and the client window will receieve
the event like normal.

This fixes the bug where you first would need to click on a window, then
after focusing would need to perform the click again to have the client
register it.

Signed-off-by: Chris Frank <chris@cfrank.org>